user-microservice/requirements.txt
asgiref==3.5.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
google-auth==2.6.6
idna==3.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
